Split `__init__.py` into multiple modules.
